requirements.txt
pyyaml
yamlpath
requests